GC can now trace pointers in an EXTERNALSEXP
This change is necessary for us to support "dispatch tables" in RIR,
because we will have pointers to R objects in an EXTERNALSEXP.

Specifically, the first part of the EXTERNALSEXP is opaque to the GC,
and is followed by an array of R pointers. The GC must be able to
trace those pointers. For example:

    typedef struct DispatchTable {
        unsigned magic; // opaque
        size_t length;  // opaque
        SEXP entry[];   // R pointers
    } DispatchTable;

To access the pointer array, we take a pointer to the beginning of
the struct, move it to the end of the struct, and then move it
*back* so it points to the beginning of the array.

We use the TRUELENGTH field to indicate the length of the array.

This change is backwards compatible, as all existing EXTERNALSEXP
have TRUELENGTH set to 0.